requirements.txt
streamlit
pandas
openpyxl
pyyaml
streamlit-authenticator
reportlab
